+++ tests/testthat/test_ts_bind.R
@@ -88,6 +88,7 @@ test_that("ts_bind works with scalars", {
 })
 
 test_that("ts_bind works with short series and scalars (#197)", {
-  ts_bind(ts_tbl(mdeaths)[1:1,], 1)
+  ans <- ts_bind(ts_tbl(mdeaths)[1:1,], 1)
+  expect_is(ans, "tbl_df")
 })