+++ tests/testthat/test_issues.R
@@ -15,3 +15,8 @@ test_that("regular zoo and xts can be processed", {
   expect_is(ts_tbl(z), "tbl")
   expect_is(ts_tbl(xts::as.xts(z)), "tbl")
 })
+
+test_that("ensure time is always ordered (#202)", {
+  unorderd <- ts_tbl(ts(1:3, start = 2000))[c(2, 3, 1), ]
+  expect_true(all(diff(ts_tbl(unorderd)$time) > 0))
+})