+++ NEWS.md
@@ -4,6 +4,7 @@
 ## Features
 
 - export ts_first_of_period #178
+- regular zoo and xts can be processed #175
 
 ## Bug fixes