+++ NEWS.md
@@ -12,6 +12,7 @@
 - ts to data frame conversion works with slight offsets #186
 - ts_trend does retransform #193
 - tslist: Use ids to name single elements if possible #194
+- ts_bind works with short series and scalars #197
 
 
 # tsbox 0.2.1 (2020-04-29)