+++ NEWS.md
@@ -6,6 +6,7 @@
 - export ts_first_of_period #178
 - regular zoo and xts can be processed #175, #189
 - ts_na_interpolation() exported and added to examples in ?ts_examples #127
+- point to tempdisagg::td() for low to high frequency conversion #142
 
 ## Bug fixes