+++ NEWS.md
@@ -7,6 +7,7 @@
 - regular zoo and xts can be processed #175, #189
 - ts_na_interpolation() exported and added to examples in ?ts_examples #127
 - point to tempdisagg::td() for low to high frequency conversion #142
+- better error messages if series contain duplicates #181
 
 ## Bug fixes