+++ NEWS.md
@@ -1,16 +1,24 @@
 # tsbox 0.2.1.9000 (2021-??-??)
 
-
 ## Features
 
 - export ts_first_of_period #178
 - regular zoo and xts can be processed #175, #189
+- ts_na_interpolation() exported and added to examples in ?ts_examples #127
 
 ## Bug fixes
 
 - ts to data frame conversion works with slight offsets #186
 - ts_trend does retransform #193
 - tslist: Use ids to name single elements if possible #194
+- ts_bind works with short series and scalars #197
+
+## Documentation
+
+- point to tempdisagg::td() for low to high frequency conversion #142
+- better error messages if series contain duplicates #181
+- better error messages when converting one row data frames #179
+- better error messages when time col contains NA #184
 
 
 # tsbox 0.2.1 (2020-04-29)