+++ NEWS.md
@@ -12,6 +12,7 @@
 - ts_trend does retransform #193
 - tslist: Use ids to name single elements if possible #194
 - ts_bind works with short series and scalars #197
+- more tolerance towards misspecified ts objects #195
 
 ## Documentation